requirements.txt
# Core dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Scraping
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
feedparser==6.0.11
playwright==1.41.2

# Content extraction & cleaning
newspaper3k==0.2.8
python-dateutil==2.8.2

# Sentiment Analysis
transformers==4.37.2
torch==2.2.0
sentencepiece==0.1.99

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.1
schedule==1.2.1
loguru==0.7.2
